Es power

+++ bin/public/installSonar.py
@@ -15,7 +15,7 @@ __status__ = "Production"
 
 import os
 
-from general import x
+from general import x,download_file
 from scopen import scOpen
 import app
 import config
@@ -27,7 +27,7 @@ import version
 SCRIPT_VERSION = 1
 
 def build_commands(commands):
-    commands.add("install-sonar",   install_sonar,   help="Install Sonar server.")
+    commands.add("install-sonar",   install_sonar,   help="Install Sonar server. version")
     commands.add("uninstall-sonar", uninstall_sonar, help="Uninstall Sonar server.")
 
 
@@ -36,15 +36,17 @@ def install_sonar(args):
     Install and configure sonar on the local host.
 
     '''
+    if (len(args) != 2):
+      raise Exception("syco install-espower Logstash Version [syco install-sonare 5.0]")
     if os.path.isfile("/bin/java"):
       '''
       Is Java installed ?
       '''
       x('yum install wget unzip -y')
-      os.system('wget http://dist.sonar.codehaus.org/sonarqube-5.0.zip')
-      x('mv sonarqube-5.0.zip /opt/sonarqube-5.0.zip')
-      x('unzip /opt/sonarqube-5.0.zip')
-      x('ln -s /opt/sonarqube-5.0 /opt/sonarqube')
+      x('rm -rf /opt/sonarqube')
+      download_file('http://dist.sonar.codehaus.org/sonarqube-{0}.zip'.format(args[1]))
+      x('mv /opt/syco/installtemp/sonarqube-{0}.zip /opt/sonarqube.zip'.format(args[1]))
+      x('unzip /opt/sonarqube.zip')
       x('/opt/sonarqube/bin/linux-x86-64/sonar.sh start')
       x('rm /opt/sonarqube-5.0.zip')
     else:
